requirements.txt
pytest==6.0.1
pytest-cov==2.10.0
Faker==4.0.3
mock==4.0.2
pyjwt==1.7.1
flask==1.1.2
gunicorn==20.0.4
flask-cors==3.0.8
flasgger==0.9.5
sentry-sdk[flask]==0.17.3
celery==4.4.2
boto3==1.12.44
mailchimp3==3.0.13
mandrill==1.0.59
requests==2.25.0

# Wrappers
bcrypt==3.1.7
pymongo==3.11.0
